Make method update in LocalMovieGateway return the number of movies updated

+++ app/src/main/java/com/josecognizant/popmovies/domain/model/LocalMovieGateway.java
@@ -11,7 +11,13 @@ public interface LocalMovieGateway {
 
     void persist(List<Movie> moviesToPersist);
 
-    long update(Movie movieToUpdate);
+    /**
+     * Updates a movie in the local database
+     *
+     * @param movieToUpdate Movie to be updated
+     * @return Number of rows updated
+     */
+    int update(Movie movieToUpdate);
 
     void delete(Movie movieToDelete);
 }
